+++ CHANGELOG.md
@@ -37,6 +37,9 @@
     - Renamed `block::Text::to_builder()` to `into_builder()`.
     - Renamed `all_is_cubes::color_block!()` to `all_is_cubes::block::from_color!()`.
 
+- `all-is-cubes-mesh` library:
+    - Renamed `GfxVertex` to `Vertex`.
+
 - `all-is-cubes-ui` library:
     - `apps::Session::settings()` replaces `graphics_options_mut()`.